refactor: replace UserWord
